// src/castor/protos/node_models.ts
export namespace io.iohk.atala.prism.protos {
// ...
    export class CompressedECKeyData extends pb_1.Message {
        #one_of_decls: number[][] = [];
        constructor(data?: any[] | {
            curve?: string;
            data?: Uint8Array;
        }) {
            super();
            pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
            if (!Array.isArray(data) && typeof data == "object") {
                if ("curve" in data && data.curve != undefined) {
                    this.curve = data.curve;
                }
                if ("data" in data && data.data != undefined) {
                    this.data = data.data;
                }
            }
        }
        get curve() {
            return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
        }
        set curve(value: string) {
            pb_1.Message.setField(this, 1, value);
        }
        get data() {
            return pb_1.Message.getFieldWithDefault(this, 2, new Uint8Array(0)) as Uint8Array;
        }
        set data(value: Uint8Array) {
            pb_1.Message.setField(this, 2, value);
        }
        static fromObject(data: {
            curve?: string;
            data?: Uint8Array;
        }): CompressedECKeyData {
            const message = new CompressedECKeyData({});
            if (data.curve != null) {
                message.curve = data.curve;
            }
            if (data.data != null) {
                message.data = data.data;
            }
            return message;
        }
        toObject() {
            const data: {
                curve?: string;
                data?: Uint8Array;
            } = {};
            if (this.curve != null) {
                data.curve = this.curve;
            }
            if (this.data != null) {
                data.data = this.data;
            }
            return data;
        }
        serialize(): Uint8Array;
        serialize(w: pb_1.BinaryWriter): void;
        serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
            const writer = w || new pb_1.BinaryWriter();
            if (this.curve.length)
                writer.writeString(1, this.curve);
            if (this.data.length)
                writer.writeBytes(2, this.data);
            if (!w)
                return writer.getResultBuffer();
        }
        static deserialize(bytes: Uint8Array | pb_1.BinaryReader): CompressedECKeyData {
            const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new CompressedECKeyData();
            while (reader.nextField()) {
                if (reader.isEndGroup())
                    break;
                switch (reader.getFieldNumber()) {
                    case 1:
                        message.curve = reader.readString();
                        break;
                    case 2:
                        message.data = reader.readBytes();
                        break;
                    default: reader.skipField();
                }
            }
            return message;
        }
        serializeBinary(): Uint8Array {
            return this.serialize();
        }
        static deserializeBinary(bytes: Uint8Array): CompressedECKeyData {
            return CompressedECKeyData.deserialize(bytes);
        }
    }
// ...
    export class CreateDIDOperation extends pb_1.Message {
        #one_of_decls: number[][] = [];
        constructor(data?: any[] | {
            did_data?: CreateDIDOperation.DIDCreationData;
        }) {
            super();
            pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
            if (!Array.isArray(data) && typeof data == "object") {
                if ("did_data" in data && data.did_data != undefined) {
                    this.did_data = data.did_data;
                }
            }
        }
        get did_data() {
            return pb_1.Message.getWrapperField(this, CreateDIDOperation.DIDCreationData, 1) as CreateDIDOperation.DIDCreationData;
        }
        set did_data(value: CreateDIDOperation.DIDCreationData) {
            pb_1.Message.setWrapperField(this, 1, value);
        }
        get has_did_data() {
            return pb_1.Message.getField(this, 1) != null;
        }
        static fromObject(data: {
            did_data?: ReturnType<typeof CreateDIDOperation.DIDCreationData.prototype.toObject>;
        }): CreateDIDOperation {
            const message = new CreateDIDOperation({});
            if (data.did_data != null) {
                message.did_data = CreateDIDOperation.DIDCreationData.fromObject(data.did_data);
            }
            return message;
        }
        toObject() {
            const data: {
                did_data?: ReturnType<typeof CreateDIDOperation.DIDCreationData.prototype.toObject>;
            } = {};
            if (this.did_data != null) {
                data.did_data = this.did_data.toObject();
            }
            return data;
        }
        serialize(): Uint8Array;
        serialize(w: pb_1.BinaryWriter): void;
        serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
            const writer = w || new pb_1.BinaryWriter();
            if (this.has_did_data)
                writer.writeMessage(1, this.did_data, () => this.did_data.serialize(writer));
            if (!w)
                return writer.getResultBuffer();
        }
        static deserialize(bytes: Uint8Array | pb_1.BinaryReader): CreateDIDOperation {
            const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new CreateDIDOperation();
            while (reader.nextField()) {
                if (reader.isEndGroup())
                    break;
                switch (reader.getFieldNumber()) {
                    case 1:
                        reader.readMessage(message.did_data, () => message.did_data = CreateDIDOperation.DIDCreationData.deserialize(reader));
                        break;
                    default: reader.skipField();
                }
            }
            return message;
        }
        serializeBinary(): Uint8Array {
            return this.serialize();
        }
        static deserializeBinary(bytes: Uint8Array): CreateDIDOperation {
            return CreateDIDOperation.deserialize(bytes);
        }
    }
    export namespace CreateDIDOperation {
        export class DIDCreationData extends pb_1.Message {
            #one_of_decls: number[][] = [];
            constructor(data?: any[] | {
                public_keys?: PublicKey[];
                services?: Service[];
            }) {
                super();
                pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [2, 3], this.#one_of_decls);
                if (!Array.isArray(data) && typeof data == "object") {
                    if ("public_keys" in data && data.public_keys != undefined) {
                        this.public_keys = data.public_keys;
                    }
                    if ("services" in data && data.services != undefined) {
                        this.services = data.services;
                    }
                }
            }
            get public_keys() {
                return pb_1.Message.getRepeatedWrapperField(this, PublicKey, 2) as PublicKey[];
            }
            set public_keys(value: PublicKey[]) {
                pb_1.Message.setRepeatedWrapperField(this, 2, value);
            }
            get services() {
                return pb_1.Message.getRepeatedWrapperField(this, Service, 3) as Service[];
            }
            set services(value: Service[]) {
                pb_1.Message.setRepeatedWrapperField(this, 3, value);
            }
            static fromObject(data: {
                public_keys?: ReturnType<typeof PublicKey.prototype.toObject>[];
                services?: ReturnType<typeof Service.prototype.toObject>[];
            }): DIDCreationData {
                const message = new DIDCreationData({});
                if (data.public_keys != null) {
                    message.public_keys = data.public_keys.map(item => PublicKey.fromObject(item));
                }
                if (data.services != null) {
                    message.services = data.services.map(item => Service.fromObject(item));
                }
                return message;
            }
            toObject() {
                const data: {
                    public_keys?: ReturnType<typeof PublicKey.prototype.toObject>[];
                    services?: ReturnType<typeof Service.prototype.toObject>[];
                } = {};
                if (this.public_keys != null) {
                    data.public_keys = this.public_keys.map((item: PublicKey) => item.toObject());
                }
                if (this.services != null) {
                    data.services = this.services.map((item: Service) => item.toObject());
                }
                return data;
            }
            serialize(): Uint8Array;
            serialize(w: pb_1.BinaryWriter): void;
            serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
                const writer = w || new pb_1.BinaryWriter();
                if (this.public_keys.length)
                    writer.writeRepeatedMessage(2, this.public_keys, (item: PublicKey) => item.serialize(writer));
                if (this.services.length)
                    writer.writeRepeatedMessage(3, this.services, (item: Service) => item.serialize(writer));
                if (!w)
                    return writer.getResultBuffer();
            }
            static deserialize(bytes: Uint8Array | pb_1.BinaryReader): DIDCreationData {
                const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new DIDCreationData();
                while (reader.nextField()) {
                    if (reader.isEndGroup())
                        break;
                    switch (reader.getFieldNumber()) {
                        case 2:
                            reader.readMessage(message.public_keys, () => pb_1.Message.addToRepeatedWrapperField(message, 2, PublicKey.deserialize(reader), PublicKey));
                            break;
                        case 3:
                            reader.readMessage(message.services, () => pb_1.Message.addToRepeatedWrapperField(message, 3, Service.deserialize(reader), Service));
                            break;
                        default: reader.skipField();
                    }
                }
                return message;
            }
            serializeBinary(): Uint8Array {
                return this.serialize();
            }
            static deserializeBinary(bytes: Uint8Array): DIDCreationData {
                return DIDCreationData.deserialize(bytes);
            }
        }
    }
// ...
}